# openflare-server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OpenFlare Server 管理端与 Agent API 文档。",
        "title": "OpenFlare Server API",
        "contact": {},
        "version": "3.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/update/latest-release": {
            "get": {
                "security": [
                    {
                        "OpenFlareTokenAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Get latest GitHub release",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "AccessTokenAuth": {
            "description": "Agent API 使用节点专属 Agent Token 或全局 Discovery Token",
            "type": "apiKey",
            "name": "X-Agent-Token",
            "in": "header"
        },
        "OpenFlareTokenAuth": {
            "description": "管理端 API 使用登录后返回的用户 Token",
            "type": "apiKey",
            "name": "OPENFLARE_TOKEN",
            "in": "header"
        }
    }
}